LayoutTests/fast/dom/add-document-child-and-reparent-old-child-during-document-child-replacement-expected.txt

CONSOLE MESSAGE: line 9: HierarchyRequestError: The operation would yield an incorrect node tree.
CONSOLE MESSAGE: line 9: HierarchyRequestError: The operation would yield an incorrect node tree.
CONSOLE MESSAGE: line 9: HierarchyRequestError: The operation would yield an incorrect node tree.
CONSOLE MESSAGE: line 9: HierarchyRequestError: The operation would yield an incorrect node tree.
CONSOLE MESSAGE: line 9: HierarchyRequestError: The operation would yield an incorrect node tree.
PASS document.replaceChild(document.body, document.documentElement) threw exception HierarchyRequestError: The operation would yield an incorrect node tree..
